Part 1 Keyword fixes
o Fixes on some keyword usage for Part 1

+++ Projects/Shapes/Shapes/Shapes/Square.cpp
@@ -11,12 +11,12 @@ FSquare::FSquare(int InSide)
 }
 
 
-int FSquare::GetArea()
+double FSquare::GetArea()
 {
     return Side * Side;
 }
 
-int FSquare::GetPerimeter()
+double FSquare::GetPerimeter()
 {
     return 4 * Side;
 }
